+++ client/containers/Project/Interface/InterfaceList/InterfaceEditForm.js
@@ -266,7 +266,7 @@ class InterfaceEditForm extends Component {
           values.req_headers = values.req_headers || [];
           values.req_body_form = values.req_body_form || [];
           let isfile = false,
-            isHavaContentType = false;
+            isHaveContentType = false;
           if (values.req_body_type === 'form') {
             values.req_body_form.forEach(item => {
               if (item.type === 'file') {
@@ -277,10 +277,10 @@ class InterfaceEditForm extends Component {
             values.req_headers.map(item => {
               if (item.name === 'Content-Type') {
                 item.value = isfile ? 'multipart/form-data' : 'application/x-www-form-urlencoded';
-                isHavaContentType = true;
+                isHaveContentType = true;
               }
             });
-            if (isHavaContentType === false) {
+            if (isHaveContentType === false) {
               values.req_headers.unshift({
                 name: 'Content-Type',
                 value: isfile ? 'multipart/form-data' : 'application/x-www-form-urlencoded'
@@ -291,11 +291,11 @@ class InterfaceEditForm extends Component {
               ? values.req_headers.map(item => {
                   if (item.name === 'Content-Type') {
                     item.value = 'application/json';
-                    isHavaContentType = true;
+                    isHaveContentType = true;
                   }
                 })
               : [];
-            if (isHavaContentType === false) {
+            if (isHaveContentType === false) {
               values.req_headers = values.req_headers || [];
               values.req_headers.unshift({
                 name: 'Content-Type',
@@ -542,7 +542,7 @@ class InterfaceEditForm extends Component {
 
   // 处理批量导入参数
   handleBulkOk = () => {
-    let curValue = this.props.form.getFieldValue(this.state.bulkName);
+    let curValue = this.props.form.getFieldValue(this.state.bulkName)||[];
     // { name: '', required: '1', desc: '', example: '' }
     let newValue = [];
 
@@ -553,6 +553,7 @@ class InterfaceEditForm extends Component {
       newValue.push(valueItem);
     });
 
+    this.props.form.setFieldsValue({[this.state.bulkName]: newValue});
     this.setState({
       visible: false,
       bulkValue: null,
@@ -574,9 +575,11 @@ class InterfaceEditForm extends Component {
     let value = this.props.form.getFieldValue(name);
 
     let bulkValue = ``;
-    value.forEach(item => {
-      return (bulkValue += item.name ? `${item.name}:${item.example || ''}\n` : '');
-    });
+    if(value) {
+      value.forEach(item => {
+        return (bulkValue += item.name ? `${item.name}:${item.example || ''}\n` : '');
+      });
+    }
 
     this.setState({
       visible: true,
@@ -1291,8 +1294,8 @@ class InterfaceEditForm extends Component {
               {...formItemLayout}
               label={
                 <span>
-                  邮件通知&nbsp;
-                  <Tooltip title={'开启邮件通知，可在 项目设置 里修改'}>
+                  消息通知&nbsp;
+                  <Tooltip title={'开启消息通知，可在 项目设置 里修改'}>
                     <Icon type="question-circle-o" style={{ width: '10px' }} />
                   </Tooltip>
                 </span>