+++ config.js
@@ -1,6 +1,7 @@
 module.exports = {
   "port": "3000",
   "adminAccount": "admin@admin.com",
+  "timeout":120000,
   "db": {
     "connecString": process.env.connectString,
     "servername": "127.0.0.1",