+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## v1.8.8
+* 更新了 cross-request [教程](https://juejin.im/post/5e3bbd986fb9a07ce152b53d)，发布了最新的 3.1 版本 
+
+## v1.8.7
+* 因 chrome 官方下架了 yapi 扩展，整理了本地安装教程
+
+## v1.8.6
+* 优化 swagger 文档导入分类策略，优先使用根路径的 tags 做分类，避免出现特别多分类的问题
+
+## v1.8.5
+* 改善 swagger 自动导入，不再支持秒级别的 cron 表达式,默认使用 10分钟更新一次的频率
+* 修复输入空的 swagger地址 ，会发起请求的 bug
+* 优化 swagger 数据导入，不会导入空的分类，不会使用版本号作为分类名称
+* swagger 导入自动增加 basePath 到项目配置
+
+## v1.8.4
+* 修复 swagger 导入数据时，如果数据格式中缺少 in 的参数，会丢失请求参数
+* 修复当传入数据格式的 method 不规范时，容易导致获取对象为空，出现异常导致动态页面无法打开
+* 修复解决 json-schema-faker/json-schema-faker#453 问题
+
+## v1.8.3
+* 修复管理员无法看到所有分组的 bug
+
+## v1.8.2
+* 重构分组列表功能实现，大幅度优化首屏加载速度
+* 接口运行界面设置 header、query、form 的初始值为其示例值
+* 运行接口请求时支持自动预览HTML
+
+### v1.8.1
+* 优化插件【Swagger 自动同步】在添加地址时的服务端校验行为
+* 优化单个测试用例执行超时时间限制,从3秒改为10秒
+
+### v1.8.0
+* filtering interface on the server instead of client
+
+### v1.7.2
+* 支持接口路径模糊搜索，不包含 basepath
+
+### v1.7.1
+* 废弃 yapi.ymfe.org 文档站点
+
 ### v1.7.0
 * fix：修复md两个undefined以及run_auto_test中执行用例id问题 #1024